+++ src/main/java/kobe/KobeException.java
@@ -1,3 +1,5 @@
+package kobe;
+
 public class KobeException extends Exception {
 
     public KobeException(String message) {
