+++ src/main/java/kobe/command/AddCommand.java
@@ -1,3 +1,12 @@
+package kobe.command;
+
+import kobe.task.Deadline;
+import kobe.task.Event;
+import kobe.task.TaskList;
+import kobe.task.Todo;
+import kobe.util.Storage;
+import kobe.util.Ui;
+
 import java.io.*;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
