+++ src/main/java/kobe/command/Command.java
@@ -1,3 +1,9 @@
+package kobe.command;
+
+import kobe.util.Storage;
+import kobe.task.TaskList;
+import kobe.util.Ui;
+
 import java.io.*;
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws IOException;
