+++ src/main/java/kobe/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package kobe.command;
+
+import kobe.util.Storage;
+import kobe.task.TaskList;
+import kobe.util.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
