+++ src/main/java/kobe/command/MarkCommand.java
@@ -1,16 +1,23 @@
+package kobe.command;
+
+import kobe.util.Storage;
+import kobe.task.Task;
+import kobe.task.TaskList;
+import kobe.util.Ui;
+
 import java.io.IOException;
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     private final int taskIndex;
 
-    public UnmarkCommand(int taskIndex) {
+    public MarkCommand(int taskIndex) {
         this.taskIndex = taskIndex - 1; // Convert to 0-based index
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
         Task task = tasks.getTask(taskIndex);
-        task.markAsNotDone();
-        System.out.println("OK, I've marked this task as not done yet:");
+        task.markAsDone();
+        System.out.println("Nice! I've marked this task as done:");
         System.out.println("  " + task);
         storage.save(tasks.getTasks());
     }
