+++ src/main/java/kobe/task/Event.java
@@ -1,3 +1,7 @@
+package kobe.task;
+
+import kobe.task.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 public class Event extends Task {
