+++ src/main/java/kobe/task/Todo.java
@@ -1,3 +1,7 @@
+package kobe.task;
+
+import kobe.task.Task;
+
 public class Todo extends Task {
     public Todo(String name) {
         super(name);
