+++ src/main/java/kobe/util/Storage.java
@@ -1,3 +1,10 @@
+package kobe.util;
+
+import kobe.task.Deadline;
+import kobe.task.Event;
+import kobe.task.Task;
+import kobe.task.Todo;
+
 import java.io.*;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
