+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Greetings! I am Kobe Bryant.
+Greetings! I am kobe.Kobe Bryant.
  K   K   OOOOO   BBBBB   EEEEE
  K  K   O     O  B    B  E
  KKK    O     O  BBBBB   EEEE
